C89 support for macOS and shared files; improvements to CMake
